chore(build): skip jacoco by default, enable on CI
This change disables the excecution of the jacoco-maven-plugin by default and enabling it on CI.

The plugin configuration is put into a a profile coverage in the parent POM. The profile is activated on modules containing the src/test/java or target/generated-test-source folder.

The profile is triggered by existence of a flag file target/executeJacoco. This is to execute the Jacoco plugin only for modules with test sources.

A second profile coverage-not-enabled is added which defines an empty property jacoco.argLine. This is because the build would fail when this property is missing and maven-surefire-plugin is executed and refers to that property. An empty value is needed then. Defining the empty property on the properties section in the parent POM did not work, the jacococ-maven-plugin did not override the property. Leaving out would leave the placeholder, which is not valid.

The goal report-aggregate of the jacoco-maven-plugin is executed after the regular build in order to execute the report aggregation only on the rector POM.