refactor(tests): Make BatchHelper independent of ProcessEngineRule
This change introduces the interface `ProcessEngineProvider`, which is implemented by the classes `ProcessEngineRule` and `PluggableProcessEngineTest`.

The BatchHelper class hierarchy is decoupled from `ProcessEngineRule` and `PluggableProcessEngineTest` by using this interface to access the ProcessEngine and its services.

This is a preparation step for migrating test cases away from these both classes.

related to #27