fix(build): Consolidate usage of maven-source-plugin
This change puts the execution of the maven-sources-plugin into a profile `sources`, which is activated by property `with-sources`. The property is `false` by default. It is set to `true` for the `sonatype-oss-release` profile, since sources are mandatory for Maven Central deployment.

Same with javadocs: A profile `javadocs` with activation by `with-javadocs` property.

Removed redundant configuration of maven-source-plugin.

Fixes #296