refactor(build): Introduce `install.skip` property
Property `install.skip` is added to the root POM with value `false`, and used to configure the `skip` property for the maven-install-plugin.

All modules that require skipping the install plugin because they do not produce a 'main' jar override the property value. This makes the definition of the maven-install-plugin in these POMs obsolete and therefore the files smaller.

Additionally skipping install for distro/run qa modules.

fixes #342