fix(tests): Fix SqlScriptTest (#377)
Fix the test when `h2-in-memory` profile is enabled by adding property `database.type` in the profile in `database/pom.xml`.

Migrate SqlScriptTest to JUnit 5. Using `@EnabledIf` feature and check for the existence of files in target/test-classes/scripts-old, which is populated when the `check-sql` profile is active.

closes #376
related to #27