Fix NPE when running integration tests locally. (#453)
Added null check for deploymentDir in OperatonBpmRunDeploymentConfiguration#getNormalizedDeploymentDir

A NPE arises if operaton.deploymentDir is null because in OperatonBpmRunDeploymentConfiguration.getNormalizedDeploymentDir the value is accessed without a check. In a regular installation the property is set in the run.bat/run.sh scripts. But these are not used for the integration tests. Instead the default is taken from OperatonBpmRunConfiguration.operatonDeploymentConfiguration. Here the value is configured to be null.