chore: Rewriting thread behavior to best practices. (#556)
We call 'await' in a loop until we run out of time or receive the
signal. We exit immediately when receiving the signal and calculate if
there is remaining time otherwise.

The 'await' can terminate either by a call from 'signal' or by timing
out. Both can happen and are probably ok, but we should at least log the
information that a timeout occurred while thread is still running.

Adding tests to ensure it behaves as before.

relates to: #537