refactor: Upgrade Spring, SpringBoot, JakartaEE EL, improve dependency mangement & merge Spring modules (#583)
This change upgrades the Spring dependencies to Spring 6.2.3. It further resolves dependencies on Spring 5.
closes #577

The internal-dependencies BOM imports now the BOMs from Spring & JakartaEE. This makes it obsolete for modules that use this BOM to declare any dependency versions for Spring or JakartaEE libraries. The internal-dependencies BOM has been cleaned up to declare any specific Spring or JakartaEE library.

As a side implication is forced to upgrade JUEL to JakartaEE EL API 5.0.0, which comes by the removal of the explicit 4.0.0 version in internal-dependencies and by using the JakartEE BOM for JakartaEE 10.0.0.
closes #579 closes #252

The SpringBoot version has been lifted to 3.4.3, which uses Spring 6.2.3 again.
closes #581

Hibernate is upgraded to version 6.6.10.
closes #584

The unused dependency camel-spring has been removed.
closes #582

Wildfly is upgraded to 35.0.1
relates to #595

Any usage of Spring 5 in dependency declarations have been removed and replaced by the version managed by the internal-dependencies BOM. This implied a change in BpmPlatformXmlLocationTest.java to avoid the usage of SimpleNamingContext, which is not available in spring-test 6.x.
closes #315

SpringProcessApplicationElResolver and SpringServletProcessApplication have been refactored to use the JakartaEE servlet API. This will enable the merge of modules engine-spring/core and engine-spring/core-6 to a single module as part of the JakartaEE migration.

Further, this PR makes the module engine-spring/core-6 obsolete. The module engine-spring/core is now directly JakartaEE dependent and free of JavaEE dependencies. Thus the modules have been merged and dependencies on engine-spring/core-6 replaced by engine-spring/core.

Shading of the engine module has been reduced. The originally packaged JakartaEL library caused conflicts on Wildfly.

The change also includes merging of the module engine-cdi/jakarta into engine-cdi/core.
related to #601