chore(build): Integrate execution of integration tests in nightly build (#587)
This change add a shell script .devenv/scripts/build/build-and-run-integration-tests.sh that eases the execution of integration tests. The shell script allows to specify by parameters to execute a certain test-suite on a specified container and database.

The script is integrated into the "Integration Build" workflow (renamed from Nightly Build), so that integration tests are executed on daily schedule.

Adds integration tests for the "Operaton" distro.