chore: Suppress warnings for unused parameters
All methods with higher than private accessibility have to be considered for now as potential API.
This implies that their signatures can't change now.

When such methods declare unused parameters then this results in warninggs that can't be resolved yet.
Therefore these methods are annotated with

  @SuppressWarnings("unused")

Such API has to be revised when API will be refactored.