feat(build): Build & deploy Javadoc
This change is adding a helper shell script `javadoc.sh` to create the Operaton Javadocs.

The script is invoked by the `reports-and-docs.yml` to create nightly Javadoc builds.

Note the handling of the target directory: The target folder for Javadocs is `javadoc/operaton/{major}.{minor}`. The Javadocs are produced into a subfolder `apidocs`, which is removed by the script.

Changes to /pom.xml are done to do the required configuration outside of the `sonatype-oss-release` profile.